tools/requirements.txt
pytest
Pillow==5.2.0
